Break out lights from Vulkan vertex shader
