+++ UWP/GPU_UWP/GPU_UWP.vcxproj
@@ -382,6 +382,8 @@
     <ClInclude Include="..\..\GPU\Common\DepalettizeShaderCommon.h" />
     <ClInclude Include="..\..\GPU\Common\DrawEngineCommon.h" />
     <ClInclude Include="..\..\GPU\Common\FramebufferManagerCommon.h" />
+    <ClInclude Include="..\..\GPU\Common\FragmentShaderGeneratorCommon.h" />
+    <ClInclude Include="..\..\GPU\Common\VertexShaderGeneratorCommon.h" />
     <ClInclude Include="..\..\GPU\Common\PresentationCommon.h" />
     <ClInclude Include="..\..\GPU\Common\GPUDebugInterface.h" />
     <ClInclude Include="..\..\GPU\Common\GPUStateUtils.h" />
@@ -441,6 +443,8 @@
     <ClCompile Include="..\..\GPU\Common\DepalettizeShaderCommon.cpp" />
     <ClCompile Include="..\..\GPU\Common\DrawEngineCommon.cpp" />
     <ClCompile Include="..\..\GPU\Common\FramebufferManagerCommon.cpp" />
+    <ClCompile Include="..\..\GPU\Common\FragmentShaderGeneratorCommon.cpp" />
+    <ClCompile Include="..\..\GPU\Common\VertexShaderGeneratorCommon.cpp" />
     <ClCompile Include="..\..\GPU\Common\PresentationCommon.cpp" />
     <ClCompile Include="..\..\GPU\Common\GPUDebugInterface.cpp" />
     <ClCompile Include="..\..\GPU\Common\GPUStateUtils.cpp" />
