+++ android/jni/Android.mk
@@ -309,6 +309,8 @@ EXEC_AND_LIB_FILES := \
   $(SRC)/GPU/GeDisasm.cpp \
   $(SRC)/GPU/Common/DepalettizeShaderCommon.cpp \
   $(SRC)/GPU/Common/FramebufferManagerCommon.cpp \
+  $(SRC)/GPU/Common/FragmentShaderGeneratorCommon.cpp \
+  $(SRC)/GPU/Common/VertexShaderGeneratorCommon.cpp \
   $(SRC)/GPU/Common/PresentationCommon.cpp \
   $(SRC)/GPU/Common/GPUDebugInterface.cpp \
   $(SRC)/GPU/Common/IndexGenerator.cpp.arm \
